Fix endless recursion by avoiding to link _delegates more than once
In the overrides to willMove(toParent parent: UIViewController?) in Detents.swift and also InteractiveDismiss.swift the controller injects itself to the chain of delegates. Apparently the function is called more than once for the same controller if the device orientation changes. When a sheet is used with both .backport.presentationDetents([.medium]) and .backport.interactiveDismissDisabled(true) then multiple willMove calls will result in a cycle in the linked delegate which will then result in an endless recursion.

To avoid the recursion and the crash we just check if _delegate has not been set before.

Resolves: #62